refactor(symbols): split handler/symbols.ts
